+++ src/main/java/homework10/FileLogger.java
@@ -1,4 +1,4 @@
-package hw10;
+package homework10;
 
 import java.io.File;
 import java.io.FileWriter;
