+++ src/main/java/homework10/LoggingLevel.java
@@ -1,4 +1,4 @@
-package hw10;
+package homework10;
 
 public enum LoggingLevel {
     INFO,
