+++ src/main/java/homework10/Main.java
@@ -1,4 +1,4 @@
-package hw10;
+package homework10;
 
 import java.io.IOException;
