+++ src/main/java/homework10/configuration.txt
@@ -1,0 +1,4 @@
+FILE- C:/hillel/hillel_homework/src/main/java/homework10/application.log
+LEVEL- DEBUG
+SIZE- 1024
+FORMAT- [%s][%s] %s%n
