--- src/main/java/hw10/configuration.txt
@@ -1,4 +1,0 @@
-FILE- C:/hillel/hillel_homework/src/main/java/hw10/application.log
-LEVEL- DEBUG
-SIZE- 1024
-FORMAT- [%s][%s] %s%n
